format

+++ contracts/CentralizedArbitrator.sol
@@ -215,7 +215,7 @@ abstract contract Arbitrator {
 contract CentralizedArbitrator is Arbitrator {
     address public owner = msg.sender;
     uint256 arbitrationPrice; // Not public because arbitrationCost already acts as an accessor.
-    uint public rulingTime;
+    uint256 public rulingTime;
     uint256 constant NOT_PAYABLE_VALUE = (2**256 - 2) / 2; // High value to be sure that the appeal is too expensive.
 
     struct DisputeStruct {
